refactor: break down large test_reports.py into smaller per-report test files
- Split 1,080+ line test_reports.py into 11 focused test files
- Created tests/reports/ directory with individual test files per report:
  * test_report_daily_sales.py
  * test_report_inventory.py
  * test_report_inventory_value.py
  * test_report_jarp.py
  * test_report_kennametal_pos.py
  * test_report_monthly_consolidation.py
  * test_report_monthly_invoices.py
  * test_report_open_orders.py
  * test_report_open_po.py
  * test_report_integration.py
- Added comprehensive README.md documenting the new structure
- Preserved all 32 existing tests with no functionality changes
- Improved maintainability and readability of test suite
- Tests can now be run individually or in parallel per report
