+++ tests/reports/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for P21 API report implementations."""
